ㅠ

+++ ModuDeal/src/main/java/com/application/moduDeal/product/dao/ProductDAO.java
@@ -26,7 +26,7 @@ public interface ProductDAO {
 
     public void deleteProductImageById(int imageId);
     
-    public void deleteProductById(Long productId);
+    public void deleteProductById(int productId);
     
     public List<Map<String, Object>> getTopLikedProducts();
